use float instead of double

+++ include/core/buffer/depthbuffer.h
@@ -14,9 +14,9 @@ namespace Core
             DepthBuffer() = delete;
             ~DepthBuffer();
 
-            double GetDepth(int x, int y) const;                            // Get Depth
-            void SetDepth(int x, int y, double depth);                      // Set Depth
-            void Clear(double default_depth = 1.0);                         // Clear buffer
+            float GetDepth(int x, int y) const;                            // Get Depth
+            void SetDepth(int x, int y, float depth);                      // Set Depth
+            void Clear(float default_depth = 1.0);                         // Clear buffer
 
             int GetWidth() const;                                           // Get buffer width
             int GetHeight() const;                                          // Get buffer height
@@ -27,7 +27,7 @@ namespace Core
         private:
             int width;
             int height;
-            std::vector<double> data;                                       // store the value of depth
+            std::vector<float> data;                                       // store the value of depth
         };
     }
 }
