+++ include/core/math/matrix.h
@@ -20,11 +20,11 @@ namespace Core
         struct Matrix4x4
         {
         public:
-            double element[4][4];
+            float element[4][4];
             
             Matrix4x4();
             Matrix4x4(const Matrix4x4& mat);
-            Matrix4x4(const double matValue[4][4]);
+            Matrix4x4(const float matValue[4][4]);
             ~Matrix4x4();
 
             // Operator Overloading
@@ -32,7 +32,7 @@ namespace Core
             Matrix4x4 operator- (const Matrix4x4& mat) const;                 // Matrix-Matrix Substract
             Matrix4x4 operator* (const Matrix4x4& mat) const;                 // Matrix-Matrix Multiply
             Vector4 operator* (const Vector4& vec) const;                     // Matrix-Vector Multiply
-            Matrix4x4 operator* (const double k) const;                       // Matrix-Constant Multiply
+            Matrix4x4 operator* (const float k) const;                       // Matrix-Constant Multiply
             Matrix4x4& operator= (const Matrix4x4& other);
 
             // Get Row/Column Vector
@@ -46,7 +46,7 @@ namespace Core
             // Affine Transformation
             static Matrix4x4 Translation(const Vector3& vec);           
             static Matrix4x4 Scale(const Vector3& vec);
-            static Matrix4x4 Rotation(double angle, _3D_Cartesian_Coord axis);
+            static Matrix4x4 Rotation(float angle, _3D_Cartesian_Coord axis);
 
             // Debug Tool
             void PrintMat() const;
